Change the size of runtime tls_index to match GOT reality.
BUG=http://code.google.com/p/nativeclient/issues/detail?id=2200
TEST=See http://code.google.com/p/nativeclient/issues/detail?id=2199

Review URL: http://codereview.chromium.org/7782008

+++ sysdeps/x86_64/dl-tls.h
@@ -21,8 +21,9 @@
 /* Type used for the representation of TLS information in the GOT.  */
 typedef struct dl_tls_index
 {
-  unsigned long int ti_module;
-  unsigned long int ti_offset;
+  /* In Native Client GOT entries are 8 bytes in size.  Hence, long long.  */
+  unsigned long long ti_module;
+  unsigned long long ti_offset;
 } tls_index;
